app/requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
python-docx
Pillow
python-multipart